move recode_as_category funcs to generator
